requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
distro==1.9.0
Django==5.2
django-allauth==65.7.0
dnspython==2.7.0
dotenv==0.9.9
fastjsonschema==2.21.1
FireWorks==2.0.4
fireworks-ai==0.15.13
Flask==3.1.0
flask-paginate==2024.4.12
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
PyJWT
MarkupSafe==3.0.2
monty==2025.3.3
nbformat==5.10.4
numpy==2.2.5
oauthlib==3.2.2
openai==1.77.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
proto-plus==1.26.1
protobuf
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tabulate==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
wsproto==1.2.0
anthropic
google-generativeai
google-cloud-pubsub==2.22.0
beautifulsoup4==4.12.2
gspread==5.12.4
oauth2client==4.1.3
